Set up placeholders for tests
