+++ tools/generator/test/PBXProductType+ExtensionsTests.swift
@@ -1,0 +1,10 @@
+import XcodeProj
+import XCTest
+
+@testable import generator
+
+class PBXProductTypeExtensionsTests: XCTestCase {
+    func testSomething() {
+        XCTFail("IMPLEMENT ME!")
+    }
+}
