+++ tools/generator/test/XCSchemeEnvironmentVariablesExtensionsTests.swift
@@ -1,0 +1,10 @@
+import XcodeProj
+import XCTest
+
+@testable import generator
+
+class XCSchemeEnvironmentVariablesExtensionsTests: XCTestCase {
+    func testSomething() {
+        XCTFail("IMPLEMENT ME!")
+    }
+}
